Fix: CCLifetimeManager Neg Pos Chunk Alignment
Resolved an issue in ChunkClusterLifetimeManager where negative
coordinates incorrectly snapped toward zero due to integer division. For
example, a position of -5 would incorrectly snap to 0 instead of -16
(given a chunk length of 16). The fix ensures floating-point division is
used before applying MathF.Floor, resulting in the correct downward snap
for negative values.

+++ Framework/ChunkClusterLifetimeManager.cs
@@ -50,9 +50,9 @@ public class ChunkClusterLifetimeManager
 
     private Vector3D<int> ChunkByPos(Vector3D<int> pos) => new
         (
-            (int)MathF.Floor(pos.X / chunkLength) * chunkLength,
-            (int)MathF.Floor(pos.Y / chunkLength) * chunkLength,
-            (int)MathF.Floor(pos.Z / chunkLength) * chunkLength
+            (int)MathF.Floor((float)pos.X / chunkLength) * chunkLength,
+            (int)MathF.Floor((float)pos.Y / chunkLength) * chunkLength,
+            (int)MathF.Floor((float)pos.Z / chunkLength) * chunkLength
         );
 
     private void ProcessRegistry(CancelationCondition cancelationCondition)
